Compliance Updates and LuminaireType Convert Added
